frontend/requirements.txt
streamlit
requests
python-socketio